v2.7.18 - Fixed ToolTips for Labels
Fixed proper ToolTips for labels. What was broken: The ToolTips for labels added ToolTips based only on function name. The function name sometimes occurred in the tag like in 'ThisPC'. 'ThisPC' is a function but 'ThisPC' is also a tag in 'OpenFileExplorerTo'. Labels for ToolTips is based on Function name and its Tag: "OpenFileExplorerTo -ThisPC". So "ThisPC" occurs 3 times when using '.Contains("ThisPC")' making a 3 bullets ToolTips for "ThisPC" function label. Fixed but may need to find another way than using the function '.Contains(x)' later.